% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_numbering}
\alias{read_numbering}
\title{Read numbering.xml in the docx}
\usage{
read_numbering(file)
}
\arguments{
\item{file}{docx file path}
}
\value{
numbering
}
\description{
Read numbering.xml in the docx
}
\examples{
\dontrun{
read_numbering("reference.docx")
}
}
